requirements.txt
albumentations==1.4.18
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.4.1.post1
torch==2.4.1
tqdm==4.66.5
